+++ packages/react/.storybook/story-config.ts
@@ -53,6 +53,7 @@ export type Stories =
   | 'Carousel'
   | 'CircularProgressAvatar'
   | 'Chip'
+  | 'CollapsibleNavbarItem'
   | 'ColorModeToggle'
   | 'Colors'
   | 'Container'
@@ -82,6 +83,7 @@ export type Stories =
   | 'MenuItem'
   | 'UserDropdownMenu'
   | 'Navbar'
+  | 'NavbarItem'
   | 'OutlinedInput'
   | 'PhoneNumberInput'
   | 'Select'
@@ -181,6 +183,9 @@ const StoryConfig: StorybookConfig = {
   Chip: {
     hierarchy: `${StorybookCategories.DataDisplay}/Chip`,
   },
+  CollapsibleNavbarItem: {
+    hierarchy: `${StorybookCategories.Navigation}/Collapsible Navbar Item`,
+  },
   ColorModeToggle: {
     hierarchy: `${StorybookCategories.Theme}/Color Mode Toggle`,
   },
@@ -254,6 +259,9 @@ const StoryConfig: StorybookConfig = {
   },
   Navbar: {
     hierarchy: `${StorybookCategories.Navigation}/Navbar`,
+  },
+  NavbarItem: {
+    hierarchy: `${StorybookCategories.Navigation}/Navbar Item`,
   },
   OutlinedInput: {
     hierarchy: `${StorybookCategories.Inputs}/Outlined Input`,